+++ src/blocks/cake.cpp
@@ -32,7 +32,7 @@
 #include "cake.h"
 
 
-bool BlockCake::onPlace(User* user, int16_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction)
+bool BlockCake::onPlace(User* user, int16_t newblock, int32_t x, int16_t y, int32_t z, int map, int8_t direction)
 {
   uint8_t oldblock;
   uint8_t oldmeta;
@@ -74,7 +74,7 @@ bool BlockCake::onPlace(User* user, int16_t newblock, int32_t x, int8_t y, int32
   return false;
 }
 
-bool BlockCake::onInteract(User* user, int32_t x, int8_t y, int32_t z, int map)
+bool BlockCake::onInteract(User* user, int32_t x, int16_t y, int32_t z, int map)
 {
   uint8_t block;
   uint8_t metadata;
@@ -105,15 +105,15 @@ bool BlockCake::onInteract(User* user, int32_t x, int8_t y, int32_t z, int map)
   }
   return false;
 }
-void BlockCake::onStartedDigging(User* user, int8_t status, int32_t x, int8_t y, int32_t z, int map, int8_t direction)
+void BlockCake::onStartedDigging(User* user, int8_t status, int32_t x, int16_t y, int32_t z, int map, int8_t direction)
 {
 }
 
-void BlockCake::onDigging(User* user, int8_t status, int32_t x, int8_t y, int32_t z, int map, int8_t direction)
+void BlockCake::onDigging(User* user, int8_t status, int32_t x, int16_t y, int32_t z, int map, int8_t direction)
 {
 }
 
-bool BlockCake::onBroken(User* user, int8_t status, int32_t x, int8_t y, int32_t z, int map,  int8_t direction)
+bool BlockCake::onBroken(User* user, int8_t status, int32_t x, int16_t y, int32_t z, int map,  int8_t direction)
 {
   uint8_t block;
   uint8_t meta;
